New reorganized folders structure
