+++ common/basePage.js
