+++ pages/textFormPage/inputs.js
