+++ pages/textFormPage/pageObject.js
